fix(react-zoom-clone): minor linting adjustments, move channel type to constants
