+++ sample-apps/react/zoom-clone/src/utils/constants.ts
@@ -1,1 +1,2 @@
 export const SESSION_STORAGE_KEY = 'zc:uid';
+export const DEFAULT_CHANNEL_TYPE = 'videocall';
